Attempted fix for NPE when rendering jetpack
I've seen crashes (#139?) when there's a NPE which I'm pretty sure is caused by a null Icon. If the render pass is 1, it returns ItemArmor.getIconFromDamageForRenderPass (after a few calls), which returns its overlayIcon, which is null except in the case of cloth armor.

If we return some icon, any icon, it should at least not crash.